feat(#15): Move Container component to common component
