+++ src/pages/login/login.page.tsx
@@ -1,8 +1,7 @@
 import { useAuthStateChange } from '../../hooks/auth/useAuthStateChange';
 import { useLogin } from '../../server/auth/mutations';
-import { Github } from './assets/Github';
+import { Github } from './assets/svgs/Github';
 import { Card } from './components/Card';
-import { Container } from './components/Container';
 import { LoginButton } from './components/LoginButton';
 
 export const LoginPage = () => {
@@ -10,22 +9,20 @@ export const LoginPage = () => {
   useAuthStateChange();
 
   return (
-    <Container>
-      <Card>
-        <Card.Header>Join us!</Card.Header>
+    <Card>
+      <Card.Header>Join us!</Card.Header>
 
-        <Card.Content>
-          <LoginButton onClick={() => executeLogin('github')}>
-            <Github />
-            Login with GitHub
-          </LoginButton>
+      <Card.Content>
+        <LoginButton onClick={() => executeLogin('github')}>
+          <Github />
+          Login with GitHub
+        </LoginButton>
 
-          <Card.Footer>
-            By logging in with GitHub, you agree to our Terms of Service and
-            Privacy Policy.
-          </Card.Footer>
-        </Card.Content>
-      </Card>
-    </Container>
+        <Card.Footer>
+          By logging in with GitHub, you agree to our Terms of Service and
+          Privacy Policy.
+        </Card.Footer>
+      </Card.Content>
+    </Card>
   );
 };
